DBZGoatLib updated
Formatting
C# 12 things
Var removal

Moved the DBT player detours to PostSetupContent, some weird shit with the detour.

Added Dispose/Clear cleanup for hooks and detours. Doesn't seem to say anywhere that you need to manually clean them up, but not sure what tmod does with unloaded mods; whether it completely dumps the assembly or keeps it around which would cause the dispose to not get called.

Also added Unload call for the forms here, not sure if that helped anything but they didn't seem to have an unload call when unloading the mod. Shouldn't rely on an addon to handle their unloading especially when this is the final point of unload (last mod that should be unloaded), if anything lingers its better to just remove.